LPS-54071 move to module
