+++ modules/portal/portal-jmx/src/com/liferay/portal/jmx/internal/MBeanRegistryImpl.java
@@ -12,8 +12,9 @@
  * details.
  */
 
-package com.liferay.portal.kernel.jmx;
+package com.liferay.portal.jmx.internal;
 
+import com.liferay.portal.jmx.MBeanRegistry;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
